iOS: Converted Strings to the JSON objects before sending to plugin layer + Added AFNetworking Support + Cleanup
- Profile keys are received by the plugin are in string format, earlier we were sending this string as it is to the 
  JS layer 
- Now we try to parse and convert it to the JSON object, if we are able to convert it to the JSON object, we pass the JSON otherwise
  we pass the data as it is

- Added AFNetworking library support. Earlier we were using the NSURLConnection class and we were manging threads on our own
  because of that code was complex. 
- NSURLConnectionHelper now internally uses AFNetworking to make server calls. This fixes a crash that we were having 
  on devices running iOS 7.1. 

- Did cleanup related to delegates, remove delegate code and added methods, which takes blocks as parameters
- Bumped up version number

+++ trunk/Smartphone/iOS/SlideiOSPlugin/SlideiOSPlugin/Classes/Helpers/RobotHelper/NeatoRobotHelper.h
@@ -10,7 +10,7 @@
 + (NeatoRobot *)getRobotForId:(NSString *) robotId;
 + (void)updateUserAssociatedRobots;
 + (void)updateName:(NSString *)name forRobotwithId:(NSString *)robotId;
-+ (id)setSpotDefinitionForRobotWithId:(NSString *)robotId cleaningAreaLength:(int)cleaningAreaLength cleaningAreaHeight:(int)cleaningAreaHeight;
++ (id)setSpotDefinitionForRobotWithId:(NSString *)robotId cleaningAreaLength:(NSInteger)cleaningAreaLength cleaningAreaHeight:(NSInteger)cleaningAreaHeight;
 + (id)spotDefinitionForRobotWithId:(NSString *)robotId;
 + (void)saveXMPPCallbackId:(NSString *)xmppCallbackId;
 + (NSString *)xmppCallbackId;
